requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aioredis==1.3.1
    # via -r requirements.in
aniso8601==7.0.0
    # via graphene
async-timeout==3.0.1
    # via aioredis
click==8.0.1
    # via flask
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-sockets
flask-cors==3.0.10
    # via -r requirements.in
flask-sockets==0.2.1
    # via -r requirements.in
gevent==21.1.2
    # via
    #   -r requirements.in
    #   flask-sockets
    #   gevent-websocket
gevent-websocket==0.10.1
    # via flask-sockets
graphene==2.1.9
    # via -r requirements.in
graphql-core==2.3.2
    # via
    #   graphene
    #   graphql-relay
    #   graphql-ws
graphql-relay==2.0.1
    # via graphene
graphql-ws==0.3.1
    # via -r requirements.in
greenlet==1.1.0
    # via gevent
hiredis==2.0.0
    # via aioredis
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
lorem==0.1.1
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
promise==2.3
    # via
    #   graphql-core
    #   graphql-relay
redis==3.5.3
    # via -r requirements.in
rx==1.6.1
    # via
    #   -r requirements.in
    #   graphql-core
six==1.16.0
    # via
    #   flask-cors
    #   graphene
    #   graphql-core
    #   graphql-relay
werkzeug==1.0.1
    # via
    #   -r requirements.in
    #   flask
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
